+++ activemodel/CHANGELOG.md
@@ -1,3 +1,5 @@
+*   Trim down Active Model API by removing `valid?` and `errors.full_messages` *José Valim*
+
 ## Rails 3.2.0 (January 20, 2012) ##
 
 *   Deprecated `define_attr_method` in `ActiveModel::AttributeMethods`, because this only existed to